Fix: No more temp files errors when it exists

+++ src/main.cpp
@@ -6,7 +6,7 @@
 /*   By: vdurand <vdurand@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/24 16:18:13 by antbonin          #+#    #+#             */
-/*   Updated: 2026/06/12 02:25:06 by vdurand          ###   ########.fr       */
+/*   Updated: 2026/06/12 17:25:50 by vdurand          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -163,7 +163,8 @@ static void	init_temp_directories(std::vector<Config::ServerConfig *>& servers)
 	for (size_t index = 0; index < servers.size(); index++)
 	{
 		Config::ServerConfig& server_config = *servers[index];
-			
+		if (FileSystem::exists(server_config.tmp_dir_path))
+			return ;
 		if (::mkdir(server_config.tmp_dir_path.c_str(), 0755) == -1) 
 			throw std::runtime_error(std::string("Unable to create temp directories :") + strerror(errno));
 	}
@@ -174,6 +175,8 @@ static void remove_temp_directories(std::vector<Config::ServerConfig *>& servers
 	for (size_t index = 0; index < servers.size(); index++)
 	{
 		Config::ServerConfig& server_config = *servers[index];
+		if (!FileSystem::exists(server_config.tmp_dir_path))
+			return ;
 		if (FileSystem::removeDirectoryRecursive(server_config.tmp_dir_path) == -1)
 			throw std::runtime_error("Unable to remove temp directories");
 	}
